fix: change jest file to javascript format
For an unknown reason, Typescript was issuing an error with the
'jest.config.ts' file, saying it should be under the 'src/' folder.

Since no apparent causes were found upon research, the most optimal
solution was to simply rename the file to 'jest.config.js', as by
doing so, the issue dissapears and everything keeps working as it
should.